#85: avoid modals as separate files
