+++ source/commands/bounty.js
@@ -116,7 +116,7 @@ module.exports = new CommandWrapper(customId, "Bounties are user-created objecti
 							content: "You can post a bounty for other server members to help out with. Here's some examples:\n\t• __Party Up__ Get bounty hunters to join you for a game session\n\t• __WTB/WTS__ Get the word out your looking to trade\n\t• __Achievement Get__ Get help working toward an achievement\n\nTo make a bounty, you'll need:\n\t• a title\n\t• a description\nOptionally, you can also add:\n\t• a url for an image\n\t• a start and end time (to make an event to go with your bounty)\n\nKeep in mind that while you're in charge of adding completers and ending the bounty, the bounty is still subject to server rules and moderation.",
 							components: [
 								new ActionRowBuilder().addComponents(
-									new StringSelectMenuBuilder().setCustomId("bountypostselect")
+									new StringSelectMenuBuilder().setCustomId("bountypost")
 										.setPlaceholder("XP awarded depends on slot used...")
 										.setMaxValues(1)
 										.setOptions(slotOptions)
@@ -140,7 +140,7 @@ module.exports = new CommandWrapper(customId, "Bounties are user-created objecti
 						content: "You can select one of your open bounties to edit below.\n\nKeep in mind that while you're in charge of adding completers and ending the bounty, the bounty is still subject to server rules and moderation.",
 						components: [
 							new ActionRowBuilder().addComponents(
-								new StringSelectMenuBuilder().setCustomId("bountyeditselect")
+								new StringSelectMenuBuilder().setCustomId("bountyedit")
 									.setPlaceholder("Select a bounty to edit...")
 									.setMaxValues(1)
 									.setOptions(openBounties.map(bounty => ({
