--- source/modals/_modal_blueprint.js
@@ -1,9 +1,0 @@
-const { InteractionWrapper } = require('../classes');
-
-const customId = "";
-module.exports = new InteractionWrapper(customId, 3000,
-	/** Specs */
-	(interaction, args) => {
-
-	}
-);
