+++ source/selects/_selectDictionary.js
@@ -4,13 +4,13 @@ const { InteractionWrapper } = require("../classes");
 const selectDictionary = {};
 
 for (const file of [
-	"bountyeditselect.js",
-	"bountypostselect.js",
+	"bountyedit.js",
+	"bountypost.js",
 	"bountyshowcase.js",
 	"bountyswapbounty.js",
 	"bountyswapslot.js",
 	"bountytakedown.js",
-	"evergreeneditselect.js",
+	"evergreenedit.js",
 	"evergreenshowcase.js",
 	"evergreenswapbounty.js",
 	"evergreenswapslot.js",
